+++ CHANGELOG.md
@@ -19,6 +19,8 @@ For details about compatibility between different releases, see the **Commitment
 
 ### Fixed
 
+- Next button title in the end device wizard in the Console.
+
 ### Security
 
 ## [3.10.0] - 2020-11-02